refactored files related to pet manager
